pyproject.toml
[tool.poetry]
name = "acv"
version = "0.1.0"
description = ""
authors = ["Vincent Eichhorn <vincent.eichhorn@student.hpi.de>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.12,<3.15"
torch = "^2.9.0"
torchvision = "^0.24.0"
pillow = "^12.0.0"
numpy = "^2.3.4"
fiftyone = "^1.9.0"
matplotlib = "^3.10.7"
tqdm = "^4.67.1"
umap-learn = "^0.5.9.post2"
huggingface-hub = {extras = ["cli"], version = "^0.35.3"}
transformers = "^4.57.1"
wandb = "^0.23.0"
weave = "^0.52.20"


[tool.poetry.group.dev.dependencies]
ipykernel = "^7.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
